+++ src/auth/authStrategies/facebookStrategy.ts
@@ -1,7 +1,7 @@
 export const generateFacebookStrategy = () => {
   return [
     "// Refer to https://www.npmjs.com/package/remix-auth-socials for more information",
-    'import { AuthStrategies } from "~/services/auth.server";',
+    'import { AuthStrategies } from "~/services/auth_strategies";',
     'import type { User } from "~/services/auth.server";',
     'import { FacebookStrategy } from "remix-auth-socials";',
     "",